Moved the project up one level and the release version, also added a sample sprite to use
