+++ SpriteWander.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33627.172
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SpriteWander", "SpriteWander\SpriteWander.csproj", "{966E298D-CD72-48AC-931C-A3322C1F4901}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{966E298D-CD72-48AC-931C-A3322C1F4901}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{966E298D-CD72-48AC-931C-A3322C1F4901}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{966E298D-CD72-48AC-931C-A3322C1F4901}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{966E298D-CD72-48AC-931C-A3322C1F4901}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {27C5D7E3-64EB-4ADF-9DC9-60BFCEB86660}
+	EndGlobalSection
+EndGlobal
