--- executable/SpriteWander.deps.json
@@ -1,92 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v7.0",
-    "signature": ""
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v7.0": {
-      "SpriteWander/2.0.0": {
-        "dependencies": {
-          "CommandLineParser": "2.9.1",
-          "Newtonsoft.Json": "13.0.3",
-          "OpenTK": "3.3.3",
-          "OpenTK.GLControl": "3.3.3"
-        },
-        "runtime": {
-          "SpriteWander.dll": {}
-        }
-      },
-      "CommandLineParser/2.9.1": {
-        "runtime": {
-          "lib/netstandard2.0/CommandLine.dll": {
-            "assemblyVersion": "2.9.1.0",
-            "fileVersion": "2.9.1.0"
-          }
-        }
-      },
-      "Newtonsoft.Json/13.0.3": {
-        "runtime": {
-          "lib/net6.0/Newtonsoft.Json.dll": {
-            "assemblyVersion": "13.0.0.0",
-            "fileVersion": "13.0.3.27908"
-          }
-        }
-      },
-      "OpenTK/3.3.3": {
-        "runtime": {
-          "lib/net20/OpenTK.dll": {
-            "assemblyVersion": "3.3.3.0",
-            "fileVersion": "3.3.3.0"
-          }
-        }
-      },
-      "OpenTK.GLControl/3.3.3": {
-        "dependencies": {
-          "OpenTK": "3.3.3"
-        },
-        "runtime": {
-          "lib/net20/OpenTK.GLControl.dll": {
-            "assemblyVersion": "3.3.3.0",
-            "fileVersion": "3.3.3.0"
-          }
-        }
-      }
-    }
-  },
-  "libraries": {
-    "SpriteWander/2.0.0": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    },
-    "CommandLineParser/2.9.1": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-OE0sl1/sQ37bjVsPKKtwQlWDgqaxWgtme3xZz7JssWUzg5JpMIyHgCTY9MVMxOg48fJ1AgGT3tgdH5m/kQ5xhA==",
-      "path": "commandlineparser/2.9.1",
-      "hashPath": "commandlineparser.2.9.1.nupkg.sha512"
-    },
-    "Newtonsoft.Json/13.0.3": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-HrC5BXdl00IP9zeV+0Z848QWPAoCr9P3bDEZguI+gkLcBKAOxix/tLEAAHC+UvDNPv4a2d18lOReHMOagPa+zQ==",
-      "path": "newtonsoft.json/13.0.3",
-      "hashPath": "newtonsoft.json.13.0.3.nupkg.sha512"
-    },
-    "OpenTK/3.3.3": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-KtnygHEuYj4+agNsZRtzh5iSiKlelDZvDB/OhfebEvIDX/97Pzygv8n5llZHgrI7Xa29vvkucVBeN55R6/k+4w==",
-      "path": "opentk/3.3.3",
-      "hashPath": "opentk.3.3.3.nupkg.sha512"
-    },
-    "OpenTK.GLControl/3.3.3": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-OJKQN4s7tCzdDUoLmrEEwAzywlhIu/vmLO+qhUG4FAUKwaxNqyGYbdLgzgH4RtNzP6Qs8ZDx0RxNHXyn/DPgUA==",
-      "path": "opentk.glcontrol/3.3.3",
-      "hashPath": "opentk.glcontrol.3.3.3.nupkg.sha512"
-    }
-  }
-}
